Move things to separate files in proper packages and update naming
