+++ apps/shared/src/commonMain/kotlin/com/github/kittinunf/cored/app/api/AppViewModel.kt
@@ -1,12 +1,12 @@
 package com.github.kittinunf.cored.app.api
 
-import com.github.kittinunf.cored.StoreType
+import com.github.kittinunf.cored.store.Store
 import com.github.kittinunf.cored.app.repository.UserRepositoryImpl
 import kotlinx.coroutines.CoroutineScope
 
 class AppViewModel(private val scope: CoroutineScope) {
 
-    private val store: StoreType<AppState> by lazy { AppStore(scope, UserRepositoryImpl()) }
+    private val store: Store<AppState> by lazy { AppStore(scope, UserRepositoryImpl()) }
 
     @Suppress("Unused")
     val currentState
